seperate ut, test=allcases
