+++ python/paddle/fluid/tests/unittests/test_parallel_dygraph_pipeline_parallel.py
@@ -62,14 +62,6 @@ class TestHybridPipeParallel(TestMultipleGpus):
         self.run_mnist_2gpu('hybrid_parallel_pp_clip_grad.py', eager_mode=False)
 
 
-class TestHybridPipeParallelWithVirtualStage(TestMultipleGpus):
-
-    def test_hybrid_parallel_pp_layer_with_virtual_stage(self):
-        self.run_mnist_2gpu('hybrid_parallel_pp_layer_with_virtual_stage.py')
-        self.run_mnist_2gpu('hybrid_parallel_pp_layer_with_virtual_stage.py',
-                            eager_mode=False)
-
-
 if __name__ == "__main__":
     os.environ["FLAGS_enable_eager_mode"] = "1"
     unittest.main()
